+++ gitpylib/repo.py
@@ -3,11 +3,12 @@
 # Licensed under GNU GPL, version 2.
 
 
-import common
+from . import common
 
 
 def clone(repo):
-  common.safe_git_call('clone %s .' % repo)
+  """Returns True if the clone succeeded, False if otherwise."""
+  return common.git_call('clone %s .' % repo)[0]
 
 
 def init():