+++ gitpylib/status.py
@@ -6,9 +6,8 @@
 
 
 import os
-import re
 
-import common
+from . import common
 
 
 SUCCESS = 1
@@ -47,7 +46,7 @@ def of_file(fp):
   """
   fp = common.real_case(fp)
 
-  ok, out_ls_files, unused_err = common.git_call(
+  ok, out_ls_files, _ = common.git_call(
       'ls-files -tvco --error-unmatch "%s"' % fp)
   if not ok:
     # The file doesn't exist.
@@ -79,7 +78,14 @@ def of_repo(include_tracked_unmodified_fps=True):
     for fp_under_cwd in all_fps_under_cwd:
       if fp_under_cwd not in status_codes:
         status_codes[fp_under_cwd] = None
-  for s_fp, s in status_codes.iteritems():
+  # Python 2/3 compatibility.
+  status_codes_items = []
+  try:
+    status_codes_items = status_codes.iteritems()
+  except AttributeError:
+    status_codes_items = status_codes.items()
+
+  for s_fp, s in status_codes_items:
     status = _status_from_output(s, s_fp in au_fps, s_fp)
     yield (status, s_fp)
 
@@ -92,9 +98,9 @@ def au_files(relative_to_cwd=False):
       reported. If False, all au files in the repository are reported. (Defaults
       to False.)
   """
-  out, unused_err = common.safe_git_call(
-      'ls-files -v {}'.format(
-          '--full-name "{}"'.format(
+  out, _ = common.safe_git_call(
+      'ls-files -v {0}'.format(
+          '--full-name "{0}"'.format(
               common.repo_dir()) if not relative_to_cwd else ''))
   ret = []
   # There could be dups in the output from ls-files if, for example, there are
@@ -114,13 +120,13 @@ def _is_au_file(fp):
   Args:
     fp: the filepath to check (fp must be a file not a dir).
   """
-  out, unused_err = common.safe_git_call(
-      'ls-files -v --full-name "{}"'.format(fp))
+  out, _ = common.safe_git_call(
+      'ls-files -v --full-name "{0}"'.format(fp))
   ret = False
   if out:
     f_out = common.remove_dups(out.splitlines(), lambda x: x[2:])
     if len(f_out) != 1:
-      raise Exception('Unexpected output of ls-files: {}'.format(out))
+      raise Exception('Unexpected output of ls-files: {0}'.format(out))
     ret = f_out[0][0] == 'h'
   return ret
 
@@ -145,8 +151,8 @@ def _status_porcelain(pathspec):
     # them relative to the cwd.
     return os.path.relpath(os.path.join(repo_dir, fp), cwd)
 
-  out_status, unused_err = common.safe_git_call(
-      'status --porcelain -u --ignored "{}"'.format(pathspec))
+  out_status, _ = common.safe_git_call(
+      'status --porcelain -u --ignored "{0}"'.format(pathspec))
   ret = {}
   for f_out_status in out_status.splitlines():
     # Output is in the form <status> <file path>.
@@ -181,4 +187,4 @@ def _status_from_output(s, is_au, fp):
     return ADDED_MODIFIED
   elif s == 'AA' or s == 'M ' or s == 'DD' or 'U' in s:
     return IN_CONFLICT
-  raise Exception('Failed to get status of file {}, s is "{}"'.format(fp, s))
+  raise Exception('Failed to get status of file {0}, s is "{1}"'.format(fp, s))